bot/requirements.txt
discord.py
openai
sqlalchemy
psycopg2-binary
cryptography
